feat: improve keydown event handler
