+++ packages/editor-ui/src/components/canvas/elements/edges/CanvasEdge.vue
@@ -2,7 +2,7 @@
 /* eslint-disable vue/no-multiple-template-root */
 import type { Connection, EdgeProps } from '@vue-flow/core';
 import { BaseEdge, EdgeLabelRenderer, getBezierPath } from '@vue-flow/core';
-import { computed, onMounted, onUnmounted, toRef, useCssModule } from 'vue';
+import { computed, useCssModule } from 'vue';
 import { useI18n } from '@/composables/useI18n';
 
 const emit = defineEmits<{
@@ -13,8 +13,6 @@ const props = defineProps<EdgeProps>();
 
 const i18n = useI18n();
 const $style = useCssModule();
-
-const selected = toRef(props, 'selected');
 
 const edgeStyle = computed(() => ({
 	strokeWidth: 2,
@@ -43,26 +41,8 @@ const connection = computed<Connection>(() => ({
 	targetHandle: props.targetHandleId,
 }));
 
-onMounted(() => {
-	document.addEventListener('keydown', onKeyDown);
-});
-
-onUnmounted(() => {
-	document.removeEventListener('keydown', onKeyDown);
-});
-
 function onDelete() {
 	emit('delete', connection.value);
-}
-
-function onKeyDown(e: KeyboardEvent) {
-	if (!selected.value) {
-		return;
-	}
-
-	if (e.key === 'Delete') {
-		onDelete();
-	}
 }
 </script>
