+++ packages/editor-ui/src/plugins/i18n/locales/en.json
@@ -1258,6 +1258,8 @@
 	"nodeView.zoomToFit": "Zoom to Fit",
 	"nodeView.replaceMe": "Replace Me",
 	"nodeView.setupTemplate": "Set up template",
+	"nodeViewV2.showError.editingNotAllowed": "Editing is not allowed",
+	"nodeViewV2.showError.failedToCreateNode": "Failed to create node",
 	"contextMenu.node": "node | nodes",
 	"contextMenu.sticky": "sticky note | sticky notes",
 	"contextMenu.selectAll": "Select all",
